moved tests to be closer to source
