+++ src/client/api/__tests__/iconProvider.spec.ts
@@ -1,12 +1,12 @@
 /**
  * @jest-environment jsdom
  */
-import iconData from '../../../src/client/api/icons.json';
+import iconData from '../icons.json';
 
 import IconProvider, {
   IconInfo,
   AppTag,
-} from '../../../src/client/api/iconProvider';
+} from '../iconProvider';
 const ip = IconProvider.Instance;
 const defaultType = {
   isImage: false,
