+++ src/client/utils/__tests__/orgInfo.spec.ts
@@ -7,7 +7,7 @@ import {
   getLinkedOrgs,
   lookupUserOrgs,
   linkNarrativeToOrg,
-} from '../../../src/client/utils/orgInfo';
+} from '../orgInfo';
 
 describe('Organizations API testing', () => {
   const authToken = 'someToken';
